Connected Restaurant to Buyer APIS, added the orders page functionality to restaurant

+++ RestaurantFrontend/src/main/java/com/project/RestaurantFrontend/controller/AppController.java
@@ -1,5 +1,6 @@
 package com.project.RestaurantFrontend.controller;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.project.RestaurantFrontend.dao.AppDao;
 import com.project.RestaurantFrontend.entity.Food_Dish;
+import com.project.RestaurantFrontend.entity.Orders;
 import com.project.RestaurantFrontend.entity.User;
 import com.project.RestaurantFrontend.prevalent.CurrentUser;
 
@@ -54,28 +56,32 @@ public class AppController {
 	
 	@RequestMapping("/addmenu")
 	public String menu() {
+		if(notLoggedIn())return "login";
 		return "addMenu";
 	}
 	
 	@RequestMapping("/add")
 	public String Menuadd(@ModelAttribute Food_Dish foodDish){
+		if(notLoggedIn())return "login";
 		Integer mobile_number=CurrentUser.MOBILE_NUMBER;
 		foodDish.setMobile_number(mobile_number);
-		foodDish.setId(mobile_number+foodDish.getName());
+		foodDish.setId(mobile_number+(new Date()).toString());
 		appDao.addFood(foodDish);
 		return "redirect:/food";
 	}
 	
 	@RequestMapping("/deletefood/{id}")
-	public String Menuadd(@PathVariable String id){
+	public String Menudelete(@PathVariable String id){
+		if(notLoggedIn())return "login";
 		appDao.deleteFood(id);
 		return "redirect:/food";
 	}
 	
 	@RequestMapping("/food")
 	public String food(Model m){
+		if(notLoggedIn())return "login";
 		Integer mobile_number=CurrentUser.MOBILE_NUMBER;
-		List<Food_Dish> list=appDao.getAllFood();  
+		List<Food_Dish> list=appDao.getAllFoodOfRestaurant(mobile_number);  
         m.addAttribute("list",list);
 		return "menu";
 	}
@@ -83,4 +89,18 @@ public class AppController {
 		return CurrentUser.MOBILE_NUMBER==null;
 	}
 
+	@RequestMapping(value ="/seeOrders")
+	public String seeOrders(Model model) {
+		if(notLoggedIn())return "login";
+		List<Orders> orders=appDao.getOrders();
+		model.addAttribute("orders",orders);
+		return "seeOrders";
+	}
+	
+	@RequestMapping("/markAsPickedUp/{orderId}")
+	public String markOrderAsPickedUp(@PathVariable String orderId) {
+		if(notLoggedIn())return "login";
+		appDao.markOrderAsPickedUp(orderId);
+		return "redirect:/seeOrders";
+	}
 }
